Pair internal/game test files with split source files
